==> 06_KIPRIS_TopicGen.ipynb <==
#######################################################################
<br> ######  Made by: Dr. Keungoui Kim 
<br> ######  Title: KISDI Megatrend Project - Data Prep 06. Topic Generation
<br> ######  goal : Megatrend Tech Analysis
<br> ######  Data set: KIPRIS
<br> ######  Time Span: 
<br> ######  Variables
<br> ######  Input: 
<br> ######  Output: 
<br> ######  Methodology: 
<br> ######  Time-stamp: 
<br> ######  Notice :
<br> #######################################################################

from torch import cuda

device = f'cuda:{cuda.current_device()}' if cuda.is_available() else 'cpu'

print(device)
# conda install pytorch==1.13.0 torchvision==0.14.0 torchaudio==0.13.0 pytorch-cuda=11.7 -c pytorch -c nvidia

import ollama
import pandas as pd

desiredModel = 'llama3.1:70b' # llama3.2
file_list = ['가상화-융합화','자동화-노동 형태 다변화','초개인화-맞춤화','플랫폼의 전방위적 확산']

base_prompt = ''' These words are the list of keywords describing Mega-trend derived from topic modelling analysis.
Here, Mega-trend indicates a major, long-term, and far-reaching trend that has potential to significantly impact society and economy.
Use these Mega Trend related words to generate a significant topic label that summarizes them.  
I want only a single topic label written in a following form: <Label in Korean> | <Label in English>.
Generate only a topics no extra descriptions.'''

for file in file_list:
    result_df = pd.read_csv('bertopic_results_'+str(file)+'.csv')
    result_df = result_df.loc[result_df.Topic != -1].reset_index()
    
    temp = []
    for i in range(result_df.shape[0]):
        prompt = (result_df['Top Words'][i] + base_prompt)
        response = ollama.chat(model=desiredModel,
                       messages= [{
                           'role':'user',
                           'content':prompt,}])

        OllamaResponse = response['message']['content']
        print(OllamaResponse)
        
        temp.append(OllamaResponse)
        
    result_df['label'] = temp
    result_df.to_csv('bertopic_results_label_'+str(file)+'.csv', index=False, encoding="utf-8-sig")

print(prompt)
